Switch DummyInterestDB to an singleton object too
